Fixed a few problems with updating sample counts at the Directory
** DirectoryApi.java **

Two changes here:

1. I got write permission errors with "/api/v2/eu_bbmri_eric_collections/size".
   Esther van Enckfort told me to use /api/v2/eu_bbmri_eric_DE_collections/size.
   This produced an "entity not found" error. So this is still an open
   issue.
2. "/api/v2/eu_bbmri_eric_collections?attrs=id" only gets the first 100 Collections.
   I had to add extra arguments to get all Collections.

** FhirApi.java **

I have added a number of utility functions.

** FhirReporting.java **

The primary change was in fetchCollectionSizes(), which was making a call to
a resource at fhir.bbmri.de that does not yet exist. As a stopgap, I
added some code that checks specifically for the situation that we have
with GBA/BBMRI, where there is only one Biobank and only one Collection.
In this case, all samples can be assigned to the Collection.

Some additional utility functions have also been added.

Modified to work with national nodes

The usual operation of the Directory is to take changes made to the
national nodes and use them to feed nightly updates of the European
Directory.

The accepted way for entering new data is to update the national nodes
first and then let the automated updates propagate the changes to the
European Directory.

Fortunately, all national nodes use the same host. To distinguish
between them, they use different Directory entities, containing the
relevant country code. For example, "eu_bbmri_eric_DE_biobanks" is the
entity for German node biobanks, and "eu_bbmri_eric_DE_collections" is the
entity for German node collections.

I have modified the code to make it use national node entities rather
than European entities. I have taken advantage of the fact that
Collection and Biobank IDs that conform to the BBMRI standards carry the
country code for the relevant national node.

I have also incremented the minor version number in the POM, since this
is really a fix and not a new feature.