Fixed a few problems with updating sample counts at the Directory
** DirectoryApi.java **

Two changes here:

1. I got write permission errors with "/api/v2/eu_bbmri_eric_collections/size".
   Esther van Enckfort told me to use /api/v2/eu_bbmri_eric_DE_collections/size.
   This produced an "entity not found" error. So this is still an open
   issue.
2. "/api/v2/eu_bbmri_eric_collections?attrs=id" only gets the first 100 Collections.
   I had to add extra arguments to get all Collections.

** FhirApi.java **

I have added a number of utility functions.

** FhirReporting.java **

The primary change was in fetchCollectionSizes(), which was making a call to
a resource at fhir.bbmri.de that does not yet exist. As a stopgap, I
added some code that checks specifically for the situation that we have
with GBA/BBMRI, where there is only one Biobank and only one Collection.
In this case, all samples can be assigned to the Collection.

Some additional utility functions have also been added.