Modified to work with national nodes
The usual operation of the Directory is to take changes made to the
national nodes and use them to feed nightly updates of the European
Directory.

The accepted way for entering new data is to update the national nodes
first and then let the automated updates propagate the changes to the
European Directory.

Fortunately, all national nodes use the same host. To distinguish
between them, they use different Directory entities, containing the
relevant country code. For example, "eu_bbmri_eric_DE_biobanks" is the
entity for German node biobanks, and "eu_bbmri_eric_DE_collections" is the
entity for German node collections.

I have modified the code to make it use national node entities rather
than European entities. I have taken advantage of the fact that
Collection and Biobank IDs that conform to the BBMRI standards carry the
country code for the relevant national node.

I have also incremented the minor version number in the POM, since this
is really a fix and not a new feature.