Updated to include extra attributes besides sample count
Uses the Directory API PUT to insert a range of new attributes into
the Directory with a single API call.

If a default collection ID should be used, it can either be supplied
or it can be found in the FHIR store.

The GET method uses a different data model to the PUT method, so
each one gets its own DTO.

Information obtained from the FHIR store are merged with information
obtained from the Directory in order to provide everything needed
for PUT.