Generalized the diagnosis correction
The Directory does not necessarily recognize all of the ICD 10 codes
obtained from the FHIR store and it will complain if you use codes
that it doesn't know about.

A fix for this already existed for the star model code. This fix was
generalized to make it available for the regular Direcory update code.

Basically, a map is generated that maps ICD 10 codes obtained from
FHIR onto suitable codes for the Directory. Then, all occurences
of the FHIR codes are replaced by the Directory codes.

Additionaly, some extra Javadoc was added and existing Javadoc was
cleaned up.