Changes made to allow dirsync to scale to larger datasets
I have been testing directory sync with a synthetic dataset containing
50k patients and 500k specimens. This cased a number of problems.
A specimen type that had not previously been encountered was rejected by
the Directory. There were also problems relating to the quantity
of data, which slowed down processing. I made a couple of changes
to improve efficiency. I also added lots of diagnostics, so that
it is easier to see where failures occur in the code.